feat: 지갑과 자판기에서 금액을 투입하면 선택 시간 3초 초기화
provider에서 useEffect를 사용하여 wallet, vm 컴포넌트에서 insertCoin의 sideEffect를 감지
